requirements.txt
openai
python-dotenv
llama_index
langchain
PyPDF2
PyCryptodome
gradio
black
isort